requirements.txt
aiohttp>=3.8.0
requests>=2.28.0
pymongo>=4.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
